+++ test/integration/default/serverspec/spec_helper.rb
@@ -1,0 +1,12 @@
+require 'serverspec'
+require 'pathname'
+
+include Serverspec::Helper::Exec
+include Serverspec::Helper::DetectOS
+
+RSpec.configure do |c|
+  c.before :all do
+    c.path = '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin'
+    c.os = backend(Serverspec::Commands::Base).check_os
+  end
+end